add I2S camera driver
this implements the I2S camera functionality. it has been tested using
an AI Thinker ESP32-CAM board with an OV2640 camera as well as a
Makerfabs ESP32 3.2" TFT LCD with Camera with the same camera model.
the qa-test is based on the latter board.

this connection can be found on ESP32 boards.

note that this is independent of `lcd_cam::cam` module which is a
different kind of camera.

`DataFormat::DualChannel32` is not currently available as part of this
PR since it's also not present in `esp32-camera` and it's thus unclear
what its behaviour would be and whether that should be supported.

while this is currently under `i2s::master` it should theoretically
belong to `i2s::slave` - but it relies heavily on traits which need to
be moved elsewhere so that they're accessible (currently they're under
`i2s::master::private`). this refactoring will be done at a later stage
by the maintainers :)

Co-Authored-By: Ralph Ursprung <ralph.ursprung@gmail.com>